Auto-redirect on 404 pages that end in /
This was caused by a change in jekyll leading to many of my internal
links becoming 404s... I think...

Earlier I had some logic which would give the reader a clue, but I think
it's a better user experience just to redirect. It's only going to
redirect once, so it's not that scary.