Use jekyll-feed, and misc improvements
Starting with the misc improvements:

1. Remove custom stylesheet per-post feature

I don't use it anywhere :/

2. Get rid of baseurl

I can't see any reason to keep using baseurl. I kind of forget why I
was.

3. jekyll-feed

OK, onto the big change:

jekyll-feed is nice because it's just a convenient thing for a gem to
make it rather than my own crufty file. It has some caveats:

- it has only the 10 latest posts (so be it)
- I believe it has the full posts, but not 100% sure because the Firefox
  preview only has the first paragraph. But the file definitely has the
  whole post

I'm glad to be using it though... it didn't exist (at least as part of
the github-pages bundle) when I made my blog, and I would've used it if
I could've. I want my blog to be a good example of a jekyll site, and I
think that includes not implementing a feed myself.

I'm a bit nervous that pushing this will cause a hiccup where
subscribers (if I have any besides myself) will get the latest posts
again. We'll see I guess. When that happens for sites I like, it doesn't
offend me too bad.