Generate default opengraph previews for posts
What a fun little lark this turned out to be. At first I thought someone
else's blog was doing this and I tried to figure out how, and then I
realized it wasn't actually doing this, it was just a static image, and
i was like lol damn. But wait, actually this should be doable???

This is a super basic-ass image that I'm generating, to be sure, but I
kind of like that for now (i tell myself).

I'm tempted to just commit these generated files to the repository
becauswe it'll make builds faster, but I think let's start with this...
and see if it can actually build in github actions.